Database integration start
